pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "matplotlib>=3.10.8",
    "mcp[cli]>=1.21.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.12",
]

[tool.uv.workspace]
members = [
    "Clinisight",
]
